failed_payments_message/requirements.txt
Flask>=1.1.2
gunicorn>=20.0.4
python-dotenv>=0.15.0
stripe>=2.56.0